tests/test_api_handlers.py
"""
Tests for Supervisor API handlers.
"""
import pytest
import json
from unittest.mock import Mock, AsyncMock, patch
import tornado.testing
import tornado.web

# These tests would need the actual handlers imported
# For now, creating test structure


class TestSupervisorAPIHandlers:
    """Test Supervisor REST API handlers."""

    @pytest.fixture
    def mock_supervisor(self):
        """Mock supervisor instance for API testing."""
        supervisor = Mock()
        supervisor.get_entities = Mock(return_value=[])
        supervisor.get_capabilities = Mock(return_value=[])
        supervisor.get_health = Mock(return_value={"status": "ok"})
        supervisor.deploy_capability = AsyncMock(return_value={"status": "deployed"})
        return supervisor

    @pytest.fixture
    def api_app(self, mock_supervisor):
        """Create Tornado application for testing."""
        # This would import the actual make_app function
        # from remote_services.supervisor.api.handlers import make_app
        # return make_app(mock_supervisor)
        
        # For now, return a mock app
        app = tornado.web.Application([])
        app.supervisor = mock_supervisor
        return app


class TestNodeInfoHandler:
    """Test node information endpoint."""

    @pytest.mark.asyncio
    async def test_get_node_info_success(self, mock_supervisor):
        """Test successful node info retrieval."""
        expected_info = {
            "node_id": "pi_test123",
            "hostname": "test-pi",
            "platform": "linux",
            "arch": "aarch64",
            "python_version": "3.13.5",
            "supervisor_version": "0.1.0"
        }
        
        mock_supervisor.get_node_info = Mock(return_value=expected_info)
        
        # Test would make actual HTTP request to handler
        # For now, test the mock directly
        result = mock_supervisor.get_node_info()
        assert result["node_id"] == "pi_test123"
        assert result["hostname"] == "test-pi"


class TestEntitiesHandler:
    """Test entity management endpoints."""

    @pytest.mark.asyncio
    async def test_get_entities_empty(self, mock_supervisor):
        """Test getting entities when none exist."""
        mock_supervisor.get_entities.return_value = []
        
        # Mock the API response
        result = {"entities": [], "count": 0}
        
        assert result["entities"] == []
        assert result["count"] == 0

    @pytest.mark.asyncio
    async def test_get_entities_with_data(self, mock_supervisor, sample_entity):
        """Test getting entities with data."""
        mock_entities = [sample_entity]
        mock_supervisor.get_entities.return_value = mock_entities
        
        result = {"entities": mock_entities, "count": len(mock_entities)}
        
        assert len(result["entities"]) == 1
        assert result["entities"][0]["id"] == sample_entity["id"]
        assert result["count"] == 1

    @pytest.mark.asyncio
    async def test_get_entity_state_success(self, mock_supervisor, sample_entity):
        """Test getting specific entity state."""
        mock_supervisor.get_entity_state = Mock(return_value=sample_entity)
        
        result = mock_supervisor.get_entity_state("test:sensor:temperature")
        
        assert result["id"] == "test:sensor:temperature"
        assert result["state"] == "23.5"

    @pytest.mark.asyncio
    async def test_get_entity_state_not_found(self, mock_supervisor):
        """Test getting non-existent entity state."""
        mock_supervisor.get_entity_state = Mock(return_value=None)
        
        result = mock_supervisor.get_entity_state("nonexistent:entity")
        
        assert result is None

    @pytest.mark.asyncio
    async def test_bulk_entity_states_query(self, mock_supervisor, sample_entity):
        """Test bulk entity state querying."""
        entity_ids = ["test:sensor:temperature", "test:sensor:humidity"]
        mock_states = {
            "test:sensor:temperature": sample_entity,
            "test:sensor:humidity": {**sample_entity, "id": "test:sensor:humidity"}
        }
        
        mock_supervisor.query_entity_states = Mock(return_value=mock_states)
        
        result = mock_supervisor.query_entity_states(entity_ids)
        
        assert len(result) == 2
        assert "test:sensor:temperature" in result
        assert "test:sensor:humidity" in result


class TestCapabilitiesHandler:
    """Test capability management endpoints."""

    @pytest.mark.asyncio
    async def test_get_capabilities(self, mock_supervisor):
        """Test getting list of capabilities."""
        mock_capabilities = [
            {
                "id": "network_isolation",
                "name": "Network Isolation", 
                "status": "active",
                "version": "1.0.0"
            },
            {
                "id": "ble_gatt_repeater",
                "name": "BLE GATT Repeater",
                "status": "inactive", 
                "version": "1.0.0"
            }
        ]
        
        mock_supervisor.get_capabilities.return_value = mock_capabilities
        
        result = {"capabilities": mock_capabilities, "count": len(mock_capabilities)}
        
        assert len(result["capabilities"]) == 2
        assert result["capabilities"][0]["id"] == "network_isolation"
        assert result["capabilities"][1]["status"] == "inactive"

    @pytest.mark.asyncio
    async def test_deploy_capability_success(self, mock_supervisor):
        """Test successful capability deployment."""
        config = {"enable_logging": True, "log_level": "INFO"}
        deployment_result = {"status": "deployed", "deployment_id": "deploy_123"}
        
        mock_supervisor.deploy_capability.return_value = deployment_result
        
        result = await mock_supervisor.deploy_capability("test_capability", config)
        
        assert result["status"] == "deployed"
        assert "deployment_id" in result

    @pytest.mark.asyncio  
    async def test_deploy_capability_invalid_config(self, mock_supervisor):
        """Test capability deployment with invalid config."""
        from custom_components.perimeter_control.service_descriptor import ServiceDescriptorError
        
        mock_supervisor.deploy_capability.side_effect = ServiceDescriptorError("Invalid config")
        
        with pytest.raises(ServiceDescriptorError):
            await mock_supervisor.deploy_capability("test_capability", {"invalid": "config"})

    @pytest.mark.asyncio
    async def test_execute_capability_action(self, mock_supervisor):
        """Test executing capability actions."""
        action_result = {"result": "action_executed", "message": "Success"}
        
        mock_supervisor.execute_capability_action = AsyncMock(return_value=action_result)
        
        result = await mock_supervisor.execute_capability_action(
            "photo_booth", 
            "take_photo",
            {"resolution": "1920x1080"}
        )
        
        assert result["result"] == "action_executed"
        assert result["message"] == "Success"


class TestHealthHandler:
    """Test health check endpoint."""

    @pytest.mark.asyncio
    async def test_health_check_ok(self, mock_supervisor):
        """Test health check when everything is OK."""
        health_status = {
            "status": "ok",
            "capabilities": {
                "network_isolation": "active",
                "ble_gatt_repeater": "inactive"
            },
            "uptime": 3600,
            "memory_usage": {
                "used_mb": 512,
                "total_mb": 2048
            }
        }
        
        mock_supervisor.get_health.return_value = health_status
        
        result = mock_supervisor.get_health()
        
        assert result["status"] == "ok"
        assert "capabilities" in result
        assert "uptime" in result
        assert "memory_usage" in result

    @pytest.mark.asyncio
    async def test_health_check_degraded(self, mock_supervisor):
        """Test health check when system is degraded."""
        health_status = {
            "status": "degraded",
            "capabilities": {
                "network_isolation": "error",
                "ble_gatt_repeater": "inactive"
            },
            "errors": ["Network isolation service failed"]
        }
        
        mock_supervisor.get_health.return_value = health_status
        
        result = mock_supervisor.get_health()
        
        assert result["status"] == "degraded"
        assert "errors" in result


class TestHAIntegrationHandler:
    """Test Home Assistant integration endpoints."""

    @pytest.mark.asyncio
    async def test_ha_integration_endpoint(self, mock_supervisor, sample_entity):
        """Test HA integration combined endpoint."""
        integration_data = {
            "entities": [sample_entity],
            "entity_states": {sample_entity["id"]: sample_entity},
            "services": [
                {
                    "id": "test_service",
                    "name": "Test Service",
                    "status": "active",
                    "dashboard_url": "http://localhost:8080"
                }
            ],
            "node_info": {
                "node_id": "pi_test123",
                "hostname": "test-pi",
                "supervisor_version": "0.1.0"
            },
            "config_version": "abc123",
            "timestamp": "2026-04-03T04:46:47Z"
        }
        
        mock_supervisor.get_ha_integration_data = Mock(return_value=integration_data)
        
        result = mock_supervisor.get_ha_integration_data()
        
        assert "entities" in result
        assert "services" in result
        assert "node_info" in result
        assert len(result["entities"]) == 1
        assert result["entities"][0]["id"] == sample_entity["id"]

    @pytest.mark.asyncio
    async def test_ha_dashboard_urls(self, mock_supervisor):
        """Test HA dashboard URLs endpoint."""
        dashboard_urls = {
            "photo_booth": "http://192.168.50.47:8093",
            "ble_gatt_repeater": "http://192.168.50.47:8091",
            "network_isolator": "http://192.168.50.47:5006"
        }
        
        mock_supervisor.get_dashboard_urls = Mock(return_value=dashboard_urls)
        
        result = mock_supervisor.get_dashboard_urls()
        
        assert len(result) == 3
        assert "photo_booth" in result
        assert result["photo_booth"].endswith(":8093")


class TestAPIErrorHandling:
    """Test API error handling and edge cases."""

    @pytest.mark.asyncio
    async def test_invalid_json_request(self):
        """Test handling invalid JSON in request body."""
        # This would test actual JSON parsing errors
        # Implementation depends on how handlers validate JSON
        assert True  # TODO: Implement with actual handler testing

    @pytest.mark.asyncio
    async def test_missing_required_parameters(self):
        """Test handling missing required parameters."""
        # This would test parameter validation
        assert True  # TODO: Implement with actual handler testing

    @pytest.mark.asyncio
    async def test_supervisor_exception_handling(self, mock_supervisor):
        """Test handling supervisor exceptions."""
        mock_supervisor.get_entities.side_effect = Exception("Supervisor error")
        
        # API should handle exceptions gracefully
        try:
            result = mock_supervisor.get_entities()
        except Exception as e:
            assert str(e) == "Supervisor error"

    @pytest.mark.asyncio
    async def test_concurrent_request_handling(self, mock_supervisor):
        """Test handling multiple concurrent requests."""
        # This would test concurrent request handling
        # Implementation depends on tornado configuration
        assert True  # TODO: Implement with actual load testing