tests/requirements.txt
# Testing dependencies for PerimeterControl
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-aiohttp>=1.0.4
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-timeout>=2.1.0

# Home Assistant testing
homeassistant>=2024.1.0

# HTTP and network testing  
aioresponses>=0.7.4
requests-mock>=1.10.0

# Mocking and fixtures
factory-boy>=3.2.0
freezegun>=1.2.0

# YAML and configuration testing
jsonschema>=4.17.0
yamllint>=1.29.0

# SSH testing (for deployment tests)
asyncssh>=2.22.0

# Coverage and reporting
coverage[toml]>=7.0.0
pytest-html>=3.1.0

# Development quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0